[compiler][be] Move test pragma to separate file
`Environment.ts` is getting complex so let's separate test / playground parsing logic from it
